Changed checkPermission to canView and delete

+++ src/controllers/OrdersController.php
@@ -90,9 +90,10 @@ class OrdersController extends Controller
 
         Craft::$app->getView()->registerJs('window.orderEdit = {};', View::POS_BEGIN);
         $permissions = [
-            'commerce-editOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-editOrders'),
+            'commerce-manageOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-manageOrders'),
             'commerce-deleteOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-deleteOrders'),
         ];
+
         Craft::$app->getView()->registerJs('window.orderEdit.currentUserPermissions = ' . Json::encode($permissions) . ';', View::POS_BEGIN);
         Craft::$app->getView()->registerJs('window.orderEdit.edition = "' . Plugin::getInstance()->edition . '"', View::POS_BEGIN);
 
@@ -109,7 +110,7 @@ class OrdersController extends Controller
      */
     public function actionNewOrder(): Response
     {
-        $this->requirePermission('commerce-editOrders');
+        $this->requirePermission('commerce-manageOrders');
 
         $customerId = Craft::$app->getRequest()->getParam('customerId', null);
 
@@ -145,8 +146,6 @@ class OrdersController extends Controller
      */
     public function actionEditOrder(int $orderId, Order $order = null, $paymentForm = null): Response
     {
-        $this->requirePermission('commerce-editOrders');
-
         $plugin = Plugin::getInstance();
         $variables = [];
 
@@ -157,6 +156,8 @@ class OrdersController extends Controller
                 throw new HttpException(404, Craft::t('commerce', 'Can not find order.'));
             }
         }
+
+        $this->enforceEditOrderPermissions($order);
 
         $variables['order'] = $order;
         $variables['paymentForm'] = $paymentForm;
@@ -184,7 +185,6 @@ class OrdersController extends Controller
      */
     public function actionSave(): ?Response
     {
-        $this->requirePermission('commerce-editOrders');
         $this->requirePostRequest();
 
         $data = Craft::$app->getRequest()->getBodyParam('orderData');
@@ -196,6 +196,8 @@ class OrdersController extends Controller
         if (!$order) {
             throw new HttpException(400, Craft::t('commerce', 'Invalid Order ID'));
         }
+
+        $this->enforceEditOrderPermissions($order);
 
         // Set custom field values
         $order->setFieldValuesFromRequest('fields');
@@ -248,13 +250,16 @@ class OrdersController extends Controller
     public function actionDeleteOrder(): ?Response
     {
         $this->requirePostRequest();
-        $this->requirePermission('commerce-deleteOrders');
 
         $orderId = (int)Craft::$app->getRequest()->getRequiredBodyParam('orderId');
         $order = Plugin::getInstance()->getOrders()->getOrderById($orderId);
 
         if (!$order) {
             throw new HttpException(404, Craft::t('commerce', 'Can not find order.'));
+        }
+
+        if (!$order->canDelete(Craft::$app->getUser()->getIdentity())) {
+            throw new ForbiddenHttpException('User not authorized to view this address.');
         }
 
         if (!Craft::$app->getElements()->deleteElementById($order->id)) {
@@ -272,8 +277,6 @@ class OrdersController extends Controller
      */
     public function actionRefresh(): Response
     {
-        $this->requirePermission('commerce-editOrders');
-
         $data = Craft::$app->getRequest()->getRawBody();
         $orderRequestData = Json::decodeIfJson($data);
 
@@ -282,6 +285,8 @@ class OrdersController extends Controller
         if (!$order) {
             return $this->asFailure(Craft::t('commerce', 'Invalid Order ID'));
         }
+
+        $this->enforceEditOrderPermissions($order);
 
         $this->_updateOrder($order, $orderRequestData);
 
@@ -459,7 +464,7 @@ class OrdersController extends Controller
      */
     public function actionPurchasablesTable(): Response
     {
-        $this->requirePermission('commerce-editOrders');
+        $this->requirePermission('commerce-manageOrders');
         $this->requireAcceptsJson();
 
         $request = Craft::$app->getRequest();
@@ -1002,7 +1007,7 @@ class OrdersController extends Controller
         Craft::$app->getView()->registerJs('window.orderEdit.edition = "' . Plugin::getInstance()->edition . '"', View::POS_BEGIN);
 
         $permissions = [
-            'commerce-editOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-editOrders'),
+            'commerce-manageOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-manageOrders'),
             'commerce-deleteOrders' => Craft::$app->getUser()->getIdentity()->can('commerce-deleteOrders'),
         ];
         Craft::$app->getView()->registerJs('window.orderEdit.currentUserPermissions = ' . Json::encode($permissions) . ';', View::POS_BEGIN);
@@ -1463,5 +1468,16 @@ class OrdersController extends Controller
         }
 
         return $customers;
+    }
+
+    /**
+     * @param Order $order
+     * @throws ForbiddenHttpException
+     */
+    protected function enforceEditOrderPermissions(Order $order)
+    {
+        if (!$order->canView(Craft::$app->getUser()->getIdentity())) {
+            throw new ForbiddenHttpException('User not authorized to view this address.');
+        }
     }
 }
